Reorganise functions
Second step in reorganisation.  Reorganising functions in files inside
the package according to purpose and domain.
