+++ cmd/gindoid/util.go
@@ -3,14 +3,12 @@ package main
 import (
 	"crypto/aes"
 	"crypto/cipher"
+	"crypto/md5"
 	"encoding/base64"
-	"encoding/json"
+	"encoding/hex"
 	"fmt"
-	"html/template"
 	"io/ioutil"
 	"net/http"
-	"path/filepath"
-	"time"
 
 	log "github.com/sirupsen/logrus"
 )
@@ -62,252 +60,10 @@ func IsRegisteredDOI(doi string) bool {
 	return false
 }
 
-// DoDOIJob starts the DOI registration process by authenticating with the GIN server and adding a new DOIJob to the jobQueue.
-func DoDOIJob(w http.ResponseWriter, r *http.Request, jobQueue chan DOIJob, conf *Configuration) {
-	// Make sure we can only be called with an HTTP POST request.
-	if r.Method != "POST" {
-		w.Header().Set("Allow", "POST")
-		w.WriteHeader(http.StatusMethodNotAllowed)
-		return
+func makeUUID(URI string) string {
+	if doi, ok := UUIDMap[URI]; ok {
+		return doi
 	}
-
-	dReq := DOIReq{}
-	// TODO: Error checking
-	body, _ := ioutil.ReadAll(r.Body)
-	json.Unmarshal(body, &dReq)
-	log.WithFields(log.Fields{
-		"request": fmt.Sprintf("%+v", dReq),
-		"source":  "DoDOIJob",
-	}).Debug("Received DOI request")
-
-	// verify again
-	if !verifyRequest(dReq.Repository, dReq.Username, dReq.Verification, conf.Key) {
-		log.WithFields(log.Fields{
-			"request": fmt.Sprintf("%+v", dReq),
-			"source":  "DoDOIJob",
-		}).Error("Invalid request: failed to verify")
-		dReq.Message = template.HTML(msgInvalidRequest)
-		w.WriteHeader(http.StatusUnauthorized)
-		return
-	}
-
-	user, err := conf.GIN.Session.RequestAccount(dReq.Username)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"request": fmt.Sprintf("%+v", dReq),
-			"source":  "DoDOIJob",
-			"error":   err,
-		}).Debug("Could not get userdata")
-		dReq.Message = template.HTML(msgNotLoggedIn)
-		w.WriteHeader(http.StatusUnauthorized)
-		return
-	}
-	// TODO Error checking
-	uuid := makeUUID(dReq.Repository)
-	ok, doiInfo := ValidDOIFile(dReq.Repository, conf)
-	if !ok {
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-
-	doiInfo.UUID = uuid
-	doi := conf.DOIBase + doiInfo.UUID[:6]
-	doiInfo.DOI = doi
-	dReq.DOIInfo = doiInfo
-
-	if IsRegisteredDOI(doi) {
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte(fmt.Sprintf(msgAlreadyRegistered, doi, doi)))
-		return
-	}
-	// Send email notification
-	sendMaster(&dReq, conf)
-	// Add job to queue
-	job := DOIJob{Source: dReq.Repository, User: user, Request: dReq, Name: doiInfo.UUID, Config: conf}
-	jobQueue <- job
-	// Render success
-	w.WriteHeader(http.StatusCreated)
-	w.Write([]byte(fmt.Sprintf(msgServerIsArchiving, doi)))
-}
-
-// InitDOIJob renders the page for the staging area, where information is provided to the user and offers to start the DOI registration request.
-// It validates the metadata provided from the GIN repository and shows appropriate error messages and instructions.
-func InitDOIJob(w http.ResponseWriter, r *http.Request, conf *Configuration) {
-	log.Infof("Got a new DOI request")
-	if err := r.ParseForm(); err != nil {
-		log.WithFields(log.Fields{
-			"source": "Init",
-		}).Debug("Could not parse form data")
-		w.WriteHeader(http.StatusInternalServerError)
-		// TODO: Notify via email (maybe)
-		return
-	}
-	t, err := template.ParseFiles(filepath.Join(conf.TemplatePath, "initjob.tmpl")) // Parse template file.
-	if err != nil {
-		log.WithFields(log.Fields{
-			"source": "DoDOIJob",
-			"error":  err,
-		}).Debug("Could not parse init template")
-		w.WriteHeader(http.StatusInternalServerError)
-		// TODO: Notify via email
-		return
-	}
-
-	repository := r.Form.Get("repo")
-	verification := r.Form.Get("verification")
-	username := r.Form.Get("user")
-
-	log.Infof("Got request: [repository: %s] [username: %s] [verification: %s]", repository, username, verification)
-	dReq := DOIReq{Username: username, Repository: repository, Verification: verification}
-	dReq.DOIInfo = &DOIRegInfo{}
-
-	// If all are missing, redirect to root path?
-
-	// If any of the values is missing, render invalid request page
-	if len(repository) == 0 || len(username) == 0 || len(verification) == 0 {
-		log.WithFields(log.Fields{
-			"source":       "InitDOIJob",
-			"repository":   repository,
-			"username":     username,
-			"verification": verification,
-		}).Error("Invalid request: missing fields in query string")
-		w.WriteHeader(http.StatusBadRequest)
-		dReq.Message = template.HTML(msgInvalidRequest)
-		t.Execute(w, dReq)
-		return
-	}
-
-	// Check verification string
-	if !verifyRequest(repository, username, verification, conf.Key) {
-		log.WithFields(log.Fields{
-			"source":       "InitDOIJob",
-			"repository":   repository,
-			"username":     username,
-			"verification": verification,
-		}).Error("Invalid request: failed to verify")
-		w.WriteHeader(http.StatusBadRequest)
-		dReq.Message = template.HTML(msgInvalidRequest)
-		t.Execute(w, dReq)
-		return
-	}
-
-	// check for doifile
-	if ok, doiInfo := ValidDOIFile(repository, conf); ok {
-		j, _ := json.MarshalIndent(doiInfo, "", "  ")
-		log.Debugf("Received DOI information: %s", string(j))
-		dReq.DOIInfo = doiInfo
-		err = t.Execute(w, dReq)
-		if err != nil {
-			log.WithFields(log.Fields{
-				"request": dReq,
-				"source":  "Init",
-				"error":   err,
-			}).Error("Could not parse template")
-			return
-		}
-	} else if doiInfo != nil {
-		log.WithFields(log.Fields{
-			"doiInfo": doiInfo,
-			"source":  "Init",
-			"error":   err,
-		}).Debug("DOIfile File invalid")
-		if doiInfo.Missing != nil {
-			dReq.Message = template.HTML(msgInvalidDOI + " <p>Issue:<i> " + doiInfo.Missing[0] + "</i>")
-		} else {
-			dReq.Message = template.HTML(msgInvalidDOI + msgBadEncoding)
-		}
-		dReq.DOIInfo = &DOIRegInfo{}
-		err = t.Execute(w, dReq)
-		if err != nil {
-			log.WithFields(log.Fields{
-				"doiInfo": doiInfo,
-				"request": dReq,
-				"source":  "Init",
-				"error":   err,
-			}).Error("Could not parse template")
-			return
-		}
-		return
-	} else {
-		dReq.Message = template.HTML(msgInvalidDOI)
-		t.Execute(w, dReq)
-		if err != nil {
-			log.WithFields(log.Fields{
-				"request": dReq,
-				"source":  "Init",
-				"error":   err,
-			}).Error("Could not parse template")
-			return
-		}
-		return
-	}
-}
-
-// DOIMData holds all the metadata for a dataset that's in the process of being registered.
-type DOIMData struct {
-	Data struct {
-		ID         string `json:"id"`
-		Type       string `json:"type"`
-		Attributes struct {
-			DOI        string      `json:"doi"`
-			Identifier string      `json:"identifier"`
-			URL        interface{} `json:"url"`
-			Author     []struct {
-				Literal string `json:"literal"`
-			} `json:"author"`
-			Title               string      `json:"title"`
-			ContainerTitle      string      `json:"container-title"`
-			Description         string      `json:"description"`
-			ResourceTypeSubtype string      `json:"resource-type-subtype"`
-			DataCenterID        string      `json:"data-center-id"`
-			MemberID            string      `json:"member-id"`
-			ResourceTypeID      string      `json:"resource-type-id"`
-			Version             string      `json:"version"`
-			License             interface{} `json:"license"`
-			SchemaVersion       string      `json:"schema-version"`
-			Results             []struct {
-				ID    string `json:"id"`
-				Title string `json:"title"`
-				Count int    `json:"count"`
-			} `json:"results"`
-			RelatedIdentifiers []struct {
-				RelationTypeID    string `json:"relation-type-id"`
-				RelatedIdentifier string `json:"related-identifier"`
-			} `json:"related-identifiers"`
-			Published  string      `json:"published"`
-			Registered time.Time   `json:"registered"`
-			Updated    time.Time   `json:"updated"`
-			Media      interface{} `json:"media"`
-			XML        string      `json:"xml"`
-		} `json:"attributes"`
-		Relationships struct {
-			DataCenter struct {
-				Meta struct {
-				} `json:"meta"`
-			} `json:"data-center"`
-			Member struct {
-				Meta struct {
-				} `json:"meta"`
-			} `json:"member"`
-			ResourceType struct {
-				Meta struct {
-				} `json:"meta"`
-			} `json:"resource-type"`
-		} `json:"relationships"`
-	} `json:"data"`
-}
-
-func verifyRequest(repo, username, verification, key string) bool {
-	plaintext, err := Decrypt([]byte(key), verification)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"source":       "verifyRequest",
-			"repo":         repo,
-			"username":     username,
-			"verification": verification,
-		}).Error("Invalid request: failed to decrypt verification string")
-		return false
-	}
-
-	return plaintext == repo+username
+	currMd5 := md5.Sum([]byte(URI))
+	return hex.EncodeToString(currMd5[:])
 }
